# tmf676/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |-
    ## TMF API Reference : TMF 676 - Payment

    This API provides the standardized client interface to Payment Systems for managing performed payments or refunds. Examples of Payment API originators (clients) include Web servers, mobile app servers, Contact center dashboards or retail store systems.
  title: Payment
  version: 4.0.0
servers:
- url: https://serverRootserverRoot/payment/v4/
...
paths:
  /payment:
    get:
      description: This operation list or find Payment entities
      operationId: listPayment
      parameters:
      - description: Comma-separated properties to be provided in response
        in: query
        name: fields
        schema:
          type: string
      - description: Requested index for start of resources to be provided in response
        in: query
        name: offset
        schema:
          type: integer
      - description: Requested number of resources to be provided in response
        in: query
        name: limit
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                items:
                  $ref: '#/components/schemas/Payment'
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              schema:
                type: integer
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: List or find Payment objects
      tags:
      - payment
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: payment
    post:
      description: This operation creates a Payment entity.
      operationId: createPayment
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/Payment_Create'
        description: The Payment to be created
        required: true
      responses:
        "201":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Payment'
          description: Created
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: Creates a Payment
      tags:
      - payment
      x-codegen-request-body-name: payment
      x-content-type: application/json;charset=utf-8
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: payment
  /payment/{id}:
    get:
      description: This operation retrieves a Payment entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrievePayment
      parameters:
      - description: Identifier of the Payment
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Comma-separated properties to provide in response
        in: query
        name: fields
        schema:
          type: string
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Payment'
          description: Success
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: Retrieves a Payment by ID
      tags:
      - payment
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: payment
...
components:
  schemas:
    AccountRef:
      description: Account reference. A account may be a party account or a financial
        account.
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        description: description
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of the account
          type: string
        href:
          description: Reference of the account
          type: string
        description:
          description: Detailed description of the account
          type: string
        name:
          description: Name of the account
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - id
      type: object
    ChannelRef:
      description: "The channel to which the resource reference to. e.g. channel for\
        \ selling product offerings, channel for opening a trouble ticket etc.."
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of a related entity.
          type: string
        href:
          description: Reference of the related entity.
          type: string
        name:
          description: Name of the channel.
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - id
      type: object
...
    Money:
      description: A base / value business entity used to represent money
      example:
        unit: unit
        value: 0.8008282
      properties:
        unit:
          description: Currency (ISO4217 norm uses 3 letters to define the currency)
          type: string
        value:
          description: A positive floating point number
          format: float
          type: number
      type: object
    Payment:
      description: The Payment resource represents a performed payment. It contains
        both information about the payment and the payment method used to perform
        it.
      example:
        statusDate: 2000-01-23T04:56:07.000+00:00
        amount:
          unit: unit
          value: 0.8008282
        authorizationCode: authorizationCode
        '@type': '@type'
        channel:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        description: description
        payer:
          '@referredType': '@referredType'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        totalAmount:
          unit: unit
          value: 0.8008282
        paymentItem:
        - totalAmount:
            unit: unit
            value: 0.8008282
          amount:
            unit: unit
            value: 0.8008282
          item:
            '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          '@baseType': '@baseType'
          '@type': '@type'
          id: id
          taxAmount:
            unit: unit
            value: 0.8008282
          '@schemaLocation': https://openapi-generator.tech
        - totalAmount:
            unit: unit
            value: 0.8008282
          amount:
            unit: unit
            value: 0.8008282
          item:
            '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          '@baseType': '@baseType'
          '@type': '@type'
          id: id
          taxAmount:
            unit: unit
            value: 0.8008282
          '@schemaLocation': https://openapi-generator.tech
        '@baseType': '@baseType'
        name: name
        paymentMethod:
          statusDate: 2000-01-23T04:56:07.000+00:00
          '@referredType': '@referredType'
          validFor:
            startDateTime: 2000-01-23T04:56:07.000+00:00
            endDateTime: 2000-01-23T04:56:07.000+00:00
          '@type': '@type'
          description: description
          isPreferred: true
          relatedParty:
            '@referredType': '@referredType'
            role: role
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          '@baseType': '@baseType'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
          account:
          - '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            description: description
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          - '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            description: description
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          status: status
        correlatorId: correlatorId
        id: id
        href: href
        paymentDate: 2000-01-23T04:56:07.000+00:00
        taxAmount:
          unit: unit
          value: 0.8008282
        '@schemaLocation': https://openapi-generator.tech
        account:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          description: description
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        status: status
      properties:
        id:
          description: Unique identifier of Payment
          type: string
        href:
          description: Hypertext Reference of the Payment
          type: string
        authorizationCode:
          description: Authorization code retrieved from an external payment gateway
            that could be used for conciliation
          type: string
        correlatorId:
          description: Unique identifier in the client for the payment in case it
            is needed to correlate
          type: string
        description:
          description: Text describing the contents of the payment
          type: string
        name:
          description: Screen name of the payment
          type: string
        paymentDate:
          description: Date when the payment was performed
          format: date-time
          type: string
        status:
          description: Status of the payment
          type: string
        statusDate:
          description: Date when the status was recorded
          format: date-time
          type: string
        account:
          $ref: '#/components/schemas/AccountRef'
        amount:
          $ref: '#/components/schemas/Money'
        channel:
          $ref: '#/components/schemas/ChannelRef'
        payer:
          $ref: '#/components/schemas/RelatedParty'
        paymentItem:
          items:
            $ref: '#/components/schemas/PaymentItem'
          type: array
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRefOrValue'
        taxAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
      required:
      - account
      - paymentMethod
      type: object
    Payment_Create:
      description: |-
        The Payment resource represents a performed payment. It contains both information about the payment and the payment method used to perform it.
        Skipped properties: id,href,status,statusDate,paymentDate
      properties:
        authorizationCode:
          description: Authorization code retrieved from an external payment gateway
            that could be used for conciliation
          type: string
        correlatorId:
          description: Unique identifier in the client for the payment in case it
            is needed to correlate
          type: string
        description:
          description: Text describing the contents of the payment
          type: string
        name:
          description: Screen name of the payment
          type: string
        account:
          $ref: '#/components/schemas/AccountRef'
        amount:
          $ref: '#/components/schemas/Money'
        channel:
          $ref: '#/components/schemas/ChannelRef'
        payer:
          $ref: '#/components/schemas/RelatedParty'
        paymentItem:
          items:
            $ref: '#/components/schemas/PaymentItem'
          type: array
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRefOrValue'
        taxAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
      required:
      - account
      - paymentMethod
      - totalAmount
      type: object
    PaymentItem:
      description: The paymentItem is the result of lettering process. It enables
        to assign automatically or manually part of incoming payment amount to a bill
      example:
        totalAmount:
          unit: unit
          value: 0.8008282
        amount:
          unit: unit
          value: 0.8008282
        item:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        '@baseType': '@baseType'
        '@type': '@type'
        id: id
        taxAmount:
          unit: unit
          value: 0.8008282
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of the payment Item
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        item:
          $ref: '#/components/schemas/EntityRef'
        taxAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
      required:
      - item
      type: object
...
    RelatedParty:
      description: Related Entity reference. A related party defines party or party
        role linked to a specific entity.
      example:
        '@referredType': '@referredType'
        role: role
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of a related entity.
          type: string
        href:
          description: Reference of the related entity.
          type: string
        name:
          description: Name of the related entity.
          type: string
        role:
          description: Role played by the related party
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - '@referredType'
      - id
      type: object
...
    Error:
      description: "Used when an API throws an Error, typically with a HTTP error\
        \ response-code (3xx, 4xx, 5xx)"
      example:
        reason: reason
        code: code
        '@baseType': '@baseType'
        '@type': '@type'
        message: message
        referenceError: https://openapi-generator.tech
        '@schemaLocation': https://openapi-generator.tech
        status: status
      properties:
        code:
          description: "Application relevant detail, defined in the API or a common\
            \ list."
          type: string
        reason:
          description: Explanation of the reason for the error which can be shown
            to a client user.
          type: string
        message:
          description: More details and corrective actions related to the error which
            can be shown to a client user.
          type: string
        status:
          description: HTTP Error code extension
          type: string
        referenceError:
          description: URI of documentation describing the error.
          format: uri
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class."
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name."
          type: string
      required:
      - code
      - reason
      type: object
...